src/requirements.txt
python-telegram-bot==20.5
requests==2.31.0
apimoex==1.3.0
python-dotenv==1.0.0
